feat: move enum values to dedicated files
