--- app/admin/resources/utils/link_type.ts
@@ -1,19 +1,0 @@
-import { LinkType } from "#enums/link_type";
-
-export const linkTypeEnumsValues = {
-  availableValues: [
-    { value: LinkType.Default, label: "Default" },
-    { value: LinkType.Facebook, label: "Facebook" },
-    { value: LinkType.Instagram, label: "Instagram" },
-    { value: LinkType.LinkedIn, label: "LinkedIn" },
-    { value: LinkType.Mail, label: "Email" },
-    { value: LinkType.YouTube, label: "YouTube" },
-    { value: LinkType.GitHub, label: "GitHub" },
-    { value: LinkType.TopwrBuildings, label: "TopwrBuildings" },
-    { value: LinkType.Phone, label: "Phone" },
-    { value: LinkType.X, label: "X" },
-    { value: LinkType.TikTok, label: "TikTok" },
-    { value: LinkType.Discord, label: "Discord" },
-    { value: LinkType.Twitch, label: "Twitch" },
-  ],
-};
